BackEnd/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Banco de dados
mysql-connector-python==9.1.0

# Variáveis de ambiente (.env)
python-dotenv==1.0.1

# Autenticação e hash de senha (versões estáveis)
passlib[bcrypt]==1.7.4
bcrypt==4.0.1   # ⚠️ versão compatível com Passlib e Windows

# Tokens JWT
PyJWT==2.9.0
